forecast: fix system.yaml load + reframe Tank lasts + Helsinki TZ notes
* loadYaml require: scripts/lib/yaml-load exports { load }, not the
  function directly. Without this, system.yaml didn't load in
  production → location.lat/lon were undefined → FMI requests
  returned HTTP 400 → no weather data → flat trajectory and zero
  solar charging on the card.
* "Tank lasts" now reflects the operationally meaningful threshold
  (avg drops to floor + 5 °C, when the space heater starts taking
  over) instead of the raw 12 °C floor (where stored heat is
  exhausted). Adds hoursUntilBackupNeeded to the engine output;
  client falls back to hoursUntilFloor for older API responses.
* All hour-of-day computation goes through Intl.DateTimeFormat with
  Europe/Helsinki, so the shading mask and the user-facing "Tank
  reaches floor at HH:MM" / "Collectors gain heat from HH:00 to
  HH:00" notes are correct regardless of the server's TZ env (k8s
  pods default to UTC). Previous note read "06:00 to 16:00" — that
  was UTC; it now reads "09:00 to 19:00" in EEST.
* New backup-hours note: when backup runs but the tank never crosses
  the can't-heat threshold (intermittent need), surface the heater
  hours so the €X.YZ cost figure isn't a mystery.
* Extract the empirical-fit functions and the Helsinki TZ helpers to
  server/lib/sustain-forecast-fit.js — the engine grew past the
  600-line cap with the new branches.
