+++ server/lib/sustain-forecast.js
@@ -3,11 +3,10 @@
 // sustain-forecast.js — pure 48 h tank sustain forecast engine.
 // No I/O. No npm deps.
 //
-// Timezone assumption: ts.getHours() is called in fitSolarEffectivenessByHour to group
-// readings by hour-of-day. Node.js returns LOCAL time hours, so the server timezone must
-// be set to Europe/Helsinki (UTC+2 / UTC+3) for the shading mask to align with solar noon.
-// In production the K8s pod TZ env var is set to Europe/Helsinki. In tests, synthetic
-// readings are constructed so the hour is correct regardless of test runner timezone.
+// Timezone: hour-of-day is computed via Intl.DateTimeFormat with
+// timeZone: 'Europe/Helsinki' so the shading mask and user-facing
+// hour notes are correct regardless of the server's TZ env (k8s
+// pods default to UTC).
 //
 // Exports:
 //   fitSolarEffectivenessByHour(history) → number[24]  (per-hour relative effectiveness)
@@ -16,268 +15,23 @@
 //   computeSustainForecast({ now, tankTop, tankBottom, greenhouseTemp, currentMode,
 //                            weather48h, prices48h, coefficients, config }) → forecast
 
-// ── Physical constants ──
-const TANK_THERMAL_MASS_J_PER_K = 300 * 4186;    // 300 L × 4186 J/(kg·K) = 1.2558e6 J/K
-const GH_THERMAL_MASS_J_PER_K   = 200 * 1.2 * 1005; // 200 m³ × 1.2 kg/m³ × 1005 J/(kg·K) ≈ 2.412e5 J/K
+const fit = require('./sustain-forecast-fit');
+const TANK_THERMAL_MASS_J_PER_K = fit.TANK_THERMAL_MASS_J_PER_K;
+const GH_THERMAL_MASS_J_PER_K   = fit.GH_THERMAL_MASS_J_PER_K;
+const DEFAULT_SOLAR_EFFECTIVENESS = fit.DEFAULT_SOLAR_EFFECTIVENESS;
+const helsinkiHour  = fit.helsinkiHour;
+const helsinkiHHMM  = fit.helsinkiHHMM;
+const fitSolarEffectivenessByHour = fit.fitSolarEffectivenessByHour;
+const fitEmpiricalCoefficients    = fit.fitEmpiricalCoefficients;
+
+// Engine fallbacks: used only when a caller passes a coefficients object
+// missing a field (the fit module owns the canonical defaults).
+const DEFAULT_TANK_LEAKAGE_W_PER_K = 3.0;
+const DEFAULT_GH_LOSS_W_PER_K_BASE = 25.0;
+const DEFAULT_WIND_FACTOR          = 0.05;
+
 const SECONDS_PER_HOUR = 3600;
 
-// ── Default coefficients (used when history is too sparse) ──
-const DEFAULT_TANK_LEAKAGE_W_PER_K    = 3.0;   // W/K
-const DEFAULT_GH_LOSS_W_PER_K_BASE    = 25.0;  // W/K
-// windFactor: 5% extra greenhouse loss per m/s wind speed.
-// Placeholder constant for v1 — to be empirically fit in a future version
-// once we have sufficient outdoor + greenhouse + wind history.
-const DEFAULT_WIND_FACTOR             = 0.05;
-
-const MIN_IDLE_BUCKET_MINUTES = 20;
-const MIN_BUCKETS_FOR_FIT     = 5;
-
-// ── Solar effectiveness fallback ──
-// Default flat effectiveness mask: hours 10..16 = 1.0, others = 0.
-// Used when history is too sparse to fit real shading data.
-const DEFAULT_SOLAR_EFFECTIVENESS = (function() {
-  const mask = new Array(24);
-  for (let h = 0; h < 24; h++) {
-    mask[h] = (h >= 10 && h <= 16) ? 1.0 : 0;
-  }
-  return mask;
-}());
-
-// Minimum number of rows per hour-of-day to trust the fit.
-const MIN_ROWS_PER_HOUR_FOR_SHADE = 3;
-// Minimum peak collector-outdoor excess (K) to indicate real sun was ever observed.
-const MIN_PEAK_EXCESS_K = 5;
-
-// ── Least-squares slope through origin: slope = Σ(xi·yi) / Σ(xi²) ──
-function slopeThruOrigin(xs, ys) {
-  let sumXY = 0, sumX2 = 0;
-  for (let i = 0; i < xs.length; i++) {
-    sumXY += xs[i] * ys[i];
-    sumX2 += xs[i] * xs[i];
-  }
-  return sumX2 === 0 ? null : sumXY / sumX2;
-}
-
-/**
- * Derive per-hour-of-day solar effectiveness from collector history.
- *
- * Algorithm:
- *   1. Group readings by local hour-of-day (via ts.getHours()).
- *   2. Per row: collectorExcess = collector − outdoor.
- *   3. Per hour: 80th-percentile of collectorExcess (captures sunny-day upper envelope).
- *   4. Normalise by global peak percentile → values in [0, 1].
- *   5. Clamp values < 0.1 to 0 (kills pre-sunrise warm-up noise).
- *
- * Falls back to the flat 10..16 mask when:
- *   - Any hour has fewer than MIN_ROWS_PER_HOUR_FOR_SHADE rows, OR
- *   - Global peak excess < MIN_PEAK_EXCESS_K (no real sun observed).
- *
- * @param {object} history  Same shape as fitEmpiricalCoefficients: { readings, modes }
- * @returns {number[24]}   Per-hour effectiveness in [0, 1].
- */
-function fitSolarEffectivenessByHour(history) {
-  if (!history || !Array.isArray(history.readings) || history.readings.length === 0) {
-    return DEFAULT_SOLAR_EFFECTIVENESS.slice();
-  }
-
-  const readings = history.readings;
-
-  // Group collectorExcess values by local hour-of-day.
-  const byHour = [];
-  for (let h = 0; h < 24; h++) { byHour.push([]); }
-
-  for (let i = 0; i < readings.length; i++) {
-    const r = readings[i];
-    if (typeof r.collector !== 'number' || typeof r.outdoor !== 'number') continue;
-    const ts = r.ts instanceof Date ? r.ts : new Date(r.ts);
-    const hour = ts.getHours(); // local Helsinki time (server TZ = Europe/Helsinki)
-    const excess = r.collector - r.outdoor;
-    byHour[hour].push(excess);
-  }
-
-  // Check minimum data per hour.
-  for (let h = 0; h < 24; h++) {
-    if (byHour[h].length < MIN_ROWS_PER_HOUR_FOR_SHADE) {
-      return DEFAULT_SOLAR_EFFECTIVENESS.slice();
-    }
-  }
-
-  // Compute 80th percentile per hour.
-  const percentile80 = new Array(24);
-  for (let h = 0; h < 24; h++) {
-    percentile80[h] = percentile(byHour[h], 0.80);
-  }
-
-  // Global peak.
-  let peakExcess = percentile80[0];
-  for (let h = 1; h < 24; h++) {
-    if (percentile80[h] > peakExcess) peakExcess = percentile80[h];
-  }
-
-  if (peakExcess < MIN_PEAK_EXCESS_K) {
-    return DEFAULT_SOLAR_EFFECTIVENESS.slice();
-  }
-
-  // Normalise and clamp noise.
-  const mask = new Array(24);
-  for (let h = 0; h < 24; h++) {
-    let v = percentile80[h] / peakExcess;
-    if (v < 0) v = 0;
-    if (v > 1) v = 1;
-    if (v < 0.1) v = 0;
-    mask[h] = v;
-  }
-
-  return mask;
-}
-
-/** Compute the p-th percentile of a numeric array (p in [0,1]). */
-function percentile(arr, p) {
-  if (arr.length === 0) return 0;
-  // Sort a copy.
-  const sorted = arr.slice().sort(function(a, b) { return a - b; });
-  const idx = p * (sorted.length - 1);
-  const lo  = Math.floor(idx);
-  const hi  = Math.ceil(idx);
-  if (lo === hi) return sorted[lo];
-  const frac = idx - lo;
-  return sorted[lo] * (1 - frac) + sorted[hi] * frac;
-}
-
-/**
- * Fit empirical thermal coefficients from historical sensor + mode data.
- *
- * @param {object} history
- *   history.readings  — [{ ts: Date, tankTop, tankBottom, greenhouse, outdoor, collector }] ascending
- *   history.modes     — [{ ts: Date, mode: string }] ascending
- * @returns {{ tankLeakageWPerK, greenhouseLossWPerKBase, windFactor }}
- */
-function fitEmpiricalCoefficients(history) {
-  const defaults = {
-    tankLeakageWPerK:        DEFAULT_TANK_LEAKAGE_W_PER_K,
-    greenhouseLossWPerKBase: DEFAULT_GH_LOSS_W_PER_K_BASE,
-    windFactor:              DEFAULT_WIND_FACTOR,
-    solarEffectivenessByHour: DEFAULT_SOLAR_EFFECTIVENESS.slice(),
-    usedDefaults:            true,
-  };
-
-  if (!history || !Array.isArray(history.readings) || history.readings.length < 2) {
-    return defaults;
-  }
-
-  const readings = history.readings;
-  const modes    = (history.modes || []).slice();
-
-  // Build per-reading mode labels via a forward-walking cursor.
-  const modeLabels = new Array(readings.length);
-  let cursor = 0;
-  let currentMode = 'idle';
-  // Advance past any events before the first reading.
-  while (cursor < modes.length) {
-    const tsMs = modes[cursor].ts instanceof Date
-      ? modes[cursor].ts.getTime()
-      : Number(modes[cursor].ts);
-    const r0Ms = readings[0].ts instanceof Date
-      ? readings[0].ts.getTime()
-      : Number(readings[0].ts);
-    if (tsMs <= r0Ms) { currentMode = modes[cursor].mode; cursor++; }
-    else break;
-  }
-  modeLabels[0] = currentMode;
-  for (let i = 1; i < readings.length; i++) {
-    const rMs = readings[i].ts instanceof Date
-      ? readings[i].ts.getTime()
-      : Number(readings[i].ts);
-    while (cursor < modes.length) {
-      const mMs = modes[cursor].ts instanceof Date
-        ? modes[cursor].ts.getTime()
-        : Number(modes[cursor].ts);
-      if (mMs <= rMs) { currentMode = modes[cursor].mode; cursor++; }
-      else break;
-    }
-    modeLabels[i] = currentMode;
-  }
-
-  // Bucket consecutive idle stretches: for each idle period of ≥ MIN_IDLE_BUCKET_MINUTES
-  // total duration, emit one data point per consecutive reading pair within that period.
-  // This produces multiple (deltaK, powerW) samples from a single long idle run.
-  const tankXs = [], tankYs = [];
-  const ghXs   = [], ghYs   = [];
-
-  let bucketStart = -1;
-  for (let j = 0; j <= readings.length; j++) {
-    const isIdle = j < readings.length && modeLabels[j] === 'idle';
-    if (isIdle && bucketStart === -1) {
-      bucketStart = j;
-    } else if (!isIdle && bucketStart !== -1) {
-      const bucketEnd = j - 1;
-      const r0span = readings[bucketStart];
-      const r1span = readings[bucketEnd];
-      const t0SpanMs = r0span.ts instanceof Date ? r0span.ts.getTime() : Number(r0span.ts);
-      const t1SpanMs = r1span.ts instanceof Date ? r1span.ts.getTime() : Number(r1span.ts);
-      const spanDtMin = (t1SpanMs - t0SpanMs) / 60000;
-
-      if (spanDtMin >= MIN_IDLE_BUCKET_MINUTES) {
-        // Emit one data point per consecutive pair within this idle period.
-        for (let p = bucketStart; p < bucketEnd; p++) {
-          const r0 = readings[p];
-          const r1 = readings[p + 1];
-          const t0Ms = r0.ts instanceof Date ? r0.ts.getTime() : Number(r0.ts);
-          const t1Ms = r1.ts instanceof Date ? r1.ts.getTime() : Number(r1.ts);
-          const dtMs = t1Ms - t0Ms;
-          if (dtMs <= 0) continue;
-          const dtSec = dtMs / 1000;
-          // Tank leakage sample.
-          const tankAvg0 = (r0.tankTop + r0.tankBottom) / 2;
-          const tankAvg1 = (r1.tankTop + r1.tankBottom) / 2;
-          const ghMid    = ((r0.greenhouse || 0) + (r1.greenhouse || 0)) / 2;
-          const dTankAvg = tankAvg1 - tankAvg0;
-          const dTankK   = dTankAvg / dtSec;              // K/s
-          const powerW   = dTankK * TANK_THERMAL_MASS_J_PER_K;  // W (negative = losing)
-          const deltaK   = ((tankAvg0 + tankAvg1) / 2) - ghMid; // tank-greenhouse ΔK
-          if (deltaK > 2 && isFinite(powerW) && isFinite(deltaK)) {
-            tankXs.push(deltaK);
-            tankYs.push(-powerW);  // positive = loss
-          }
-          // Greenhouse loss sample.
-          if (r0.outdoor !== undefined && r1.outdoor !== undefined &&
-              r0.greenhouse !== undefined && r1.greenhouse !== undefined) {
-            const ghAvgPair = (r0.greenhouse + r1.greenhouse) / 2;
-            const outAvg    = (r0.outdoor + r1.outdoor) / 2;
-            const dGhAvg    = r1.greenhouse - r0.greenhouse;
-            const dGhK      = dGhAvg / dtSec;
-            const ghPowerW  = dGhK * GH_THERMAL_MASS_J_PER_K;
-            const ghDeltaK  = ghAvgPair - outAvg;
-            if (ghDeltaK > 1 && isFinite(ghPowerW) && isFinite(ghDeltaK)) {
-              ghXs.push(ghDeltaK);
-              ghYs.push(-ghPowerW);
-            }
-          }
-        }
-      }
-      bucketStart = -1;
-    }
-  }
-
-  if (tankXs.length < MIN_BUCKETS_FOR_FIT && ghXs.length < MIN_BUCKETS_FOR_FIT) {
-    return defaults;
-  }
-
-  const tankSlope = tankXs.length >= MIN_BUCKETS_FOR_FIT
-    ? slopeThruOrigin(tankXs, tankYs)
-    : null;
-  const ghSlope = ghXs.length >= MIN_BUCKETS_FOR_FIT
-    ? slopeThruOrigin(ghXs, ghYs)
-    : null;
-
-  return {
-    tankLeakageWPerK:        tankSlope  !== null && tankSlope  > 0 ? tankSlope  : DEFAULT_TANK_LEAKAGE_W_PER_K,
-    greenhouseLossWPerKBase: ghSlope    !== null && ghSlope    > 0 ? ghSlope    : DEFAULT_GH_LOSS_W_PER_K_BASE,
-    windFactor:              DEFAULT_WIND_FACTOR,
-    solarEffectivenessByHour: fitSolarEffectivenessByHour(history),
-    usedDefaults:            (tankSlope === null && ghSlope === null),
-  };
-}
 
 // ── Default config ──
 const DEFAULT_CONFIG = {
@@ -345,6 +99,12 @@ function computeSustainForecast(opts) {
   let solarChargingHours     = 0;
   let greenhouseHeatingHours = 0;
   let hoursUntilFloor        = null;
+  // Hours until tank can no longer cover the greenhouse heating load alone.
+  // Always ≤ hoursUntilFloor — tank stops being able to heat the greenhouse
+  // (avg ≤ floor + 5°C) before it actually crosses the floor. This is the
+  // metric that matters operationally: it's when the space heater starts
+  // taking over, not when stored heat is fully exhausted.
+  let hoursUntilBackupNeeded = null;
   const costBreakdown          = [];
   const tankTrajectory         = [];
   const ghTrajectory           = [];
@@ -407,6 +167,7 @@ function computeSustainForecast(opts) {
         greenhouseHeatingHours += 1;
       } else {
         // Tank too depleted — space heater runs for the hour.
+        if (hoursUntilBackupNeeded === null) hoursUntilBackupNeeded = h;
         const heaterEnergyKwh = cfg.spaceHeaterKw; // 1 kW × 1 h = 1 kWh
         electricKwh += heaterEnergyKwh;
         const costEur = heaterEnergyKwh * (priceCKwh + cfg.transferFeeCKwh) / 100;
@@ -426,7 +187,7 @@ function computeSustainForecast(opts) {
     // Two gates must both pass:
     //   a) FMI radiation > threshold (there is sun in the sky at all)
     //   b) solarEffByHour[localHour] > 0 (the collectors actually see it)
-    const hourOfDay = new Date(hourMs).getHours(); // local TZ (Europe/Helsinki in prod)
+    const hourOfDay = helsinkiHour(new Date(hourMs));
     const solarEff  = solarEffByHour[hourOfDay];
     let solarGainJ = 0;
     if (radiation > cfg.solarChargeMinRadiationWm2 && solarEff > 0 && tankAvg < 60) {
@@ -499,10 +260,12 @@ function computeSustainForecast(opts) {
 
   // ── Notes ──
   const notes = buildNotes({
+    now,
     solarChargingHours,
     greenhouseHeatingHours,
     electricKwh,
     hoursUntilFloor,
+    hoursUntilBackupNeeded,
     usedDefaults,
     tankFloorC:              cfg.tankFloorC,
     weather48h:              weather,
@@ -515,6 +278,7 @@ function computeSustainForecast(opts) {
     tankTrajectory,
     greenhouseTrajectory:   ghTrajectory,
     hoursUntilFloor:        hoursUntilFloor !== null ? round2(hoursUntilFloor) : null,
+    hoursUntilBackupNeeded: hoursUntilBackupNeeded !== null ? round2(hoursUntilBackupNeeded) : null,
     electricKwh:            round4(electricKwh),
     electricCostEur:        round4(electricCostEur),
     costBreakdown,
@@ -548,13 +312,31 @@ function buildNotes(ctx) {
     if (notes.length < 3) notes.push(solNote);
   }
 
-  if (ctx.hoursUntilFloor !== null) {
-    const floorH = ctx.hoursUntilFloor;
-    const floorDate = new Date(Date.now() + floorH * 3600 * 1000);
-    const hhmm = pad2(floorDate.getUTCHours()) + ':' + pad2(floorDate.getUTCMinutes());
-    const backupH = Math.ceil(48 - floorH);
-    const floorNote = 'Tank reaches ' + ctx.tankFloorC + '°C floor at ~' + hhmm + 'Z, after which the space heater may run for up to ' + backupH + ' h.';
-    if (notes.length < 3) notes.push(floorNote);
+  // Backup-needed note: tank can heat the greenhouse for X hours, then space
+  // heater takes over. This is the operationally meaningful "tank lasts" point.
+  if (ctx.hoursUntilBackupNeeded !== null && notes.length < 3) {
+    const bH = ctx.hoursUntilBackupNeeded;
+    const switchDate = new Date(ctx.now + bH * 3600 * 1000);
+    const hhmm = helsinkiHHMM(switchDate);
+    notes.push(
+      'Tank can cover greenhouse heating for ~' + Math.round(bH) +
+      ' h (until ' + hhmm + '). After that the space heater takes over for ~' +
+      Math.round(ctx.electricKwh) + ' h.'
+    );
+  } else if (ctx.electricKwh > 0 && notes.length < 3) {
+    // Backup runs but tank never crosses the "can't heat" threshold (intermittent
+    // need, e.g. cold snaps). Still surface the cost so €X.YZ isn't a mystery.
+    notes.push(
+      'Space heater projected to run ~' + Math.round(ctx.electricKwh) +
+      ' h over the next 48 h.'
+    );
+  }
+
+  if (ctx.hoursUntilFloor !== null && notes.length < 3) {
+    const floorDate = new Date(ctx.now + ctx.hoursUntilFloor * 3600 * 1000);
+    notes.push(
+      'Tank reaches ' + ctx.tankFloorC + '°C floor at ~' + helsinkiHHMM(floorDate) + '.'
+    );
   }
 
   // Collector shading window note (only when the mask comes from real history).
